Make an extension point for plugins
Previously, the endless-sky-high-dpi plugin was bundled with the app,
and there was no way that other plugins could be installed system-wide.

To address this:

- Define an extension point, io.github.endless_sky.endless_sky.plugins,
  with the 'subdirectories' flag set. This means that an extension named
  io.github.endless_sky.endless_sky.plugins.foo will be mounted at a
  subdirectory 'foo' of the given directory.

- Define the endless-sky-high-dpi extension as one such plugin.

Further plugins could be uploaded to Flathub in the
io.github.endless_sky.endless_sky.plugins.* namespace, and advanced
users could build & install their own plugins in that namespace.

Fixes #22